[245] Move IpMreq structure to separate file.
